@@ -7,16 +7,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
 
-  cupertino_icons: ^1.0.0
-  flutter_svg: ^0.19.3
-  smooth_star_rating: ^1.1.1
+  cupertino_icons: ^1.0.3
+  flutter_svg: ^0.22.0
+  get: ^4.1.4
 
 dev_dependencies:
   flutter_test:
