@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:pizza_eshop_flutterapp/models/pizza_map.dart';
 
 class ItemImage extends StatelessWidget {
-  final PizzaMap pizzaMap;
+  final PizzaMap? pizzaMap;
 
   ItemImage({this.pizzaMap});
 
@@ -11,9 +11,9 @@ class ItemImage extends StatelessWidget {
     Size size = MediaQuery.of(context).size;
     return Center(
       child: Hero(
-        tag: pizzaMap.title,
+        tag: pizzaMap!.title!,
         child: Image.asset(
-          pizzaMap.image,
+          pizzaMap!.image!,
          // height: size.height * 0.3,
           width: size.width * 0.9,
           fit: BoxFit.fill,
