@@ -4,13 +4,13 @@ import 'package:pizza_eshop_flutterapp/utilities/constants.dart';
 
 class OrderButton extends StatelessWidget {
   const OrderButton({
-    Key key,
-    @required this.size,
+    Key? key,
+    required this.size,
     this.press,
   }) : super(key: key);
 
   final Size size;
-  final Function press;
+  final Function? press;
 
   @override
   Widget build(BuildContext context) {
@@ -24,7 +24,7 @@ class OrderButton extends StatelessWidget {
         child: Material(
           color: Colors.transparent,
           child: InkWell(
-            onTap: press,
+            onTap: press as void Function()?,
             child: Padding(
               padding: const EdgeInsets.all(20.0),
               child: Row(
