@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 
 class MainTitle extends StatelessWidget {
-  final String title;
-  MainTitle({@required this.title});
+  final String? title;
+  MainTitle({required this.title});
 
   @override
   Widget build(BuildContext context) {
     return Container(
       padding: EdgeInsets.symmetric(horizontal: 20),
       child: Text(
-        title,
+        title!,
         style: TextStyle(
           color: Colors.black,
           fontWeight: FontWeight.bold,
