@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class SubTitle extends StatelessWidget {
   final String subtitle;
 
-  SubTitle({@required this.subtitle});
+  SubTitle({required this.subtitle});
 
   @override
   Widget build(BuildContext context) {
