@@ -18,7 +18,7 @@ class HomeAppBar extends StatelessWidget {
             text: TextSpan(
               style: Theme.of(context)
                   .textTheme
-                  .headline5
+                  .headline5!
                   .copyWith(fontWeight: FontWeight.bold),
               children: [
                 TextSpan(
